move all contrib ops to one place
