+++ onnxruntime/contrib_ops/cpu_contrib_kernels.h
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "core/framework/op_kernel.h"
+#include "core/framework/kernel_registry.h"
+
+namespace onnxruntime {
+namespace contrib {
+void RegisterCpuContribKernels(KernelRegistry& kernel_registry);
+} // namespace contrib
+}  // namespace onnxruntime
