+++ onnxruntime/contrib_ops/cuda/activation/contrib_activations.cc
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "activations.h"
+#include "contrib_activations.h"
+#include "core/framework/op_kernel.h"
 
 namespace onnxruntime {
 namespace cuda {
@@ -42,16 +43,14 @@ namespace cuda {
   UNARY_ACTIVATION_OP_TYPED(name, ver, float)     \
   UNARY_ACTIVATION_OP_TYPED(name, ver, double)
 
-UNARY_ACTIVATION_OP_HFD(Elu, 6);
-UNARY_ACTIVATION_OP_HFD(HardSigmoid, 6);
-UNARY_ACTIVATION_OP_HFD(LeakyRelu, 6);
-UNARY_ACTIVATION_OP_HFD(Relu, 6);
-UNARY_ACTIVATION_OP_HFD(Selu, 6);
-UNARY_ACTIVATION_OP_HFD(Sigmoid, 6);
-UNARY_ACTIVATION_OP_HFD(Softplus, 1);
-UNARY_ACTIVATION_OP_HFD(Softsign, 1);
-UNARY_ACTIVATION_OP_HFD(Tanh, 6);
-UNARY_ACTIVATION_OP_HFD(ThresholdedRelu, 10);
+UNARY_ACTIVATION_OP_HFD(Affine, 1);
+UNARY_ACTIVATION_OP_HFD(ParametricSoftplus, 1);
+UNARY_ACTIVATION_OP_HFD(ScaledTanh, 1);
+
+
+REGISTER_ACTIVATION_KERNEL(ThresholdedRelu, 1, MLFloat16)
+REGISTER_ACTIVATION_KERNEL(ThresholdedRelu, 1, float)
+REGISTER_ACTIVATION_KERNEL(ThresholdedRelu, 1, double)
 
 }  // namespace cuda
 }  // namespace onnxruntime
