+++ onnxruntime/contrib_ops/cuda/activation/contrib_activations_impl.h
@@ -1,0 +1,32 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+#include "core/providers/cuda/activation/activations_impl.h"
+
+namespace onnxruntime {
+namespace cuda {
+
+typedef onnxruntime::cuda::CtxAlphaBeta CtxAffine;
+typedef onnxruntime::cuda::CtxAlphaBeta CtxParametricSoftplus;
+typedef onnxruntime::cuda::CtxAlphaBeta CtxScaledTanh;
+
+#define UNARY_CONTRIB_ACTIVATION_OPS() \
+  UNARY_ACTIVATION_OP_NAME(ScaledTanh) \
+  UNARY_ACTIVATION_OP_NAME(Affine)     \
+  UNARY_ACTIVATION_OP_NAME(ParametricSoftplus)
+
+#define UNARY_ACTIVATION_IMPL_DECLARATION(name) \
+  template <typename T>                         \
+  void Impl_##name(                             \
+      const T* input_data,                      \
+      T* output_data,                           \
+      const Ctx##name* func_ctx,                \
+      size_t count)
+
+#define UNARY_ACTIVATION_OP_NAME(name) UNARY_ACTIVATION_IMPL_DECLARATION(name);
+UNARY_CONTRIB_ACTIVATION_OPS()
+#undef UNARY_ACTIVATION_OP_NAME
+
+}  // namespace cuda
+}  // namespace onnxruntime
