+++ onnxruntime/contrib_ops/cuda_contrib_kernels.h
@@ -2,12 +2,12 @@
 // Licensed under the MIT License.
 
 #pragma once
-
-#include "core/framework/op_kernel.h"
 #include "core/framework/kernel_registry.h"
 
 namespace onnxruntime {
-namespace contrib {
-void RegisterContribKernels(KernelRegistry& kernel_registry);
-} // namespace contrib
+namespace cuda {
+
+void RegisterCudaContribKernels(KernelRegistry& kernel_registry);
+
+}  // namespace cuda
 }  // namespace onnxruntime
