+++ onnxruntime/core/providers/cpu/math/element_wise_ops.cc
@@ -303,15 +303,6 @@ ONNX_CPU_OPERATOR_KERNEL(
     8,
     KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
     Mean_8<float>);
-#ifndef DISABLE_CONTRIB_OPS
-namespace contrib {
-ONNX_CPU_OPERATOR_KERNEL(
-    Scale,
-    1,
-    KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
-    Scale<float>);
-}
-#endif
 
 ONNX_CPU_OPERATOR_KERNEL(
     Erf,
@@ -1014,18 +1005,6 @@ REG_EXPAND_KERNEL(uint32_t)
 REG_EXPAND_KERNEL(uint64_t)
 REG_EXPAND_KERNEL(bool)
 REG_EXPAND_KERNEL(MLFloat16)
-
-#ifndef DISABLE_CONTRIB_OPS
-namespace contrib {
-template <>
-Status Scale<float>::Compute(OpKernelContext* ctx) const {
-  auto& X = *ctx->Input<Tensor>(0);
-  auto& Y = *ctx->Output(0, X.Shape());
-  EigenMap<float>(Y) = scale_ * EigenMap<float>(X);
-  return Status::OK();
-}
-}  // namespace contrib
-#endif
 
 template <>
 Status Erf<float>::Compute(OpKernelContext* context) const {
