+++ onnxruntime/core/providers/cpu/tensor/slice.cc
@@ -32,35 +32,6 @@ ADD_TYPED_SLICE_V9_OP(MLFloat16);
 ADD_TYPED_SLICE_V9_OP(bool);
 ADD_TYPED_SLICE_V9_OP(string);
 
-#ifndef DISABLE_CONTRIB_OPS
-namespace contrib {
-#define ADD_TYPED_DYNAMIC_SLICE_OP(data_type)                                               \
-  ONNX_CPU_OPERATOR_TYPED_KERNEL(                                                           \
-      DynamicSlice,                                                                         \
-      1,                                                                                    \
-      data_type,                                                                            \
-      KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<data_type>())      \
-                        .TypeConstraint("Tind", {DataTypeImpl::GetTensorType<int32_t>(),    \
-                                                 DataTypeImpl::GetTensorType<int64_t>()}),  \
-      Slice<data_type, true>);
-
-ADD_TYPED_DYNAMIC_SLICE_OP(uint8_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(uint16_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(uint32_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(uint64_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(int8_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(int16_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(int32_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(int64_t);
-ADD_TYPED_DYNAMIC_SLICE_OP(float);
-ADD_TYPED_DYNAMIC_SLICE_OP(double);
-ADD_TYPED_DYNAMIC_SLICE_OP(MLFloat16);
-ADD_TYPED_DYNAMIC_SLICE_OP(bool);
-ADD_TYPED_DYNAMIC_SLICE_OP(string);
-
-}  // namespace contrib
-#endif
-
 #define ADD_TYPED_SLICE_V10_OP(data_type)                                                   \
   ONNX_CPU_OPERATOR_TYPED_KERNEL(                                                           \
       Slice,                                                                                \
